services/ingestion-worker/requirements.txt
anthropic==0.40.0
openai==1.54.0
feedparser==6.0.11
trafilatura==1.12.2
httpx==0.27.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyMuPDF==1.24.5
python-docx==1.1.2
